feat(nest): refresh stats on visit instead of a daily cron
Page is statically generated with a 1h revalidate, so the next visitor
after the window lapses regenerates it and re-pulls Nest; the committed
snapshot is now only the fallback for when the API is down.

Drops the midnight-UTC workflow, refresh script, nest:refresh task and
both nest API routes — a Server Component reading the lib directly is
the correct App Router shape, and only /api/subscribe needs a route.
