--- .github/workflows/nest-stats.yml
@@ -1,46 +1,0 @@
-name: Nest stats daily refresh
-
-on:
-    schedule:
-        - cron: "0 0 * * *" # midnight UTC
-    workflow_dispatch: {}
-
-concurrency:
-    group: nest-stats-refresh
-    cancel-in-progress: false
-
-permissions:
-    contents: write
-
-jobs:
-    refresh:
-        runs-on: ubuntu-latest
-        timeout-minutes: 10
-        steps:
-            - name: Checkout
-              uses: actions/checkout@v7
-              with:
-                  ref: main
-
-            # Zero-dep ESM script — Node only (no pnpm install / husky prepare).
-            - name: Setup Node
-              uses: actions/setup-node@v7
-              with:
-                  node-version: "22"
-
-            - name: Refresh Nest stats snapshot
-              run: node scripts/refresh-nest-stats.mjs
-
-            - name: Commit snapshot if changed
-              run: |
-                  git config user.name "github-actions[bot]"
-                  git config user.email "41898282+github-actions[bot]@users.noreply.github.com"
-                  git add src/data/nest-stats.json
-                  if git diff --staged --quiet; then
-                    echo "No snapshot changes"
-                    exit 0
-                  fi
-                  git commit -m "chore(nest): refresh daily stats snapshot"
-                  # Rebase onto latest main in case another push landed after checkout.
-                  git pull --rebase origin main
-                  git push origin HEAD:main
